+++ ios/MeasureSDK/SessionManager.swift
@@ -33,6 +33,7 @@ final class BaseSessionManager: SessionManager {
     private let sessionStore: SessionStore
     private let userDefaultStorage: UserDefaultStorage
     private var previousSessionCrashed = false
+    private let versionCode: String
 
     /// The current session ID.
     var sessionId: String {
@@ -49,7 +50,8 @@ final class BaseSessionManager: SessionManager {
          configProvider: ConfigProvider,
          randomizer: Randomizer = BaseRandomizer(),
          sessionStore: SessionStore,
-         userDefaultStorage: UserDefaultStorage) {
+         userDefaultStorage: UserDefaultStorage,
+         versionCode: String) {
         self.appBackgroundTimeMs = 0
         self.idProvider = idProvider
         self.logger = logger
@@ -58,6 +60,7 @@ final class BaseSessionManager: SessionManager {
         self.randomizer = randomizer
         self.sessionStore = sessionStore
         self.userDefaultStorage = userDefaultStorage
+        self.versionCode = versionCode
     }
 
     private func createNewSession() {
@@ -65,12 +68,13 @@ final class BaseSessionManager: SessionManager {
         logger.log(level: .info, message: "New session created", error: nil, data: nil)
         let session = SessionEntity(sessionId: sessionId,
                                     pid: ProcessInfo.processInfo.processIdentifier,
-                                    createdAt: Number(timeProvider.currentTimeSinceEpochInMillis),
+                                    createdAt: Number(timeProvider.now()),
                                     needsReporting: true,
                                     crashed: false)
         sessionStore.insertSession(session)
         let recentSession = RecentSession(id: session.sessionId,
-                                          createdAt: session.createdAt)
+                                          createdAt: session.createdAt,
+                                          versionCode: versionCode)
         userDefaultStorage.setRecentSession(recentSession)
     }
 
@@ -80,7 +84,7 @@ final class BaseSessionManager: SessionManager {
         }
 
         if let recentSession = userDefaultStorage.getRecentSession(), recentSession.lastEventTime != 0 {
-            let elapsedTime = timeProvider.currentTimeSinceEpochInMillis - recentSession.lastEventTime
+            let elapsedTime = timeProvider.now() - recentSession.lastEventTime
             if elapsedTime <= configProvider.sessionEndLastEventThresholdMs {
                 return recentSession.id
             }
@@ -94,15 +98,26 @@ final class BaseSessionManager: SessionManager {
         }
 
         // Check negative time to handle clock skew
-        guard (timeProvider.currentTimeSinceEpochInMillis - recentSession.createdAt) > 0 else {
+        guard (timeProvider.now() - recentSession.createdAt) > 0 else {
             return false
         }
 
-        return (timeProvider.currentTimeSinceEpochInMillis - recentSession.createdAt) >= configProvider.maxSessionDurationMs
+        return (timeProvider.now() - recentSession.createdAt) >= configProvider.maxSessionDurationMs
+    }
+
+    private func isFrameworkVersionUpdated() -> Bool {
+        if let recentSession = userDefaultStorage.getRecentSession(),
+           recentSession.versionCode == self.versionCode {
+            return false
+        }
+        return true
     }
 
     func start() {
-        if isSessonDurationThreadholdReached() {
+        if isFrameworkVersionUpdated() {
+            logger.log(level: .info, message: "Ending previous session as SDK version has been updated.", error: nil, data: nil)
+            createNewSession()
+        } else if isSessonDurationThreadholdReached() {
             logger.log(level: .info, message: "Ending previous session as maxSessionDurationMs threshold is reached.", error: nil, data: nil)
             createNewSession()
         } else if let recentSessionId = getRecentSessionId() {
@@ -114,7 +129,7 @@ final class BaseSessionManager: SessionManager {
     }
 
     func applicationDidEnterBackground() {
-        self.appBackgroundTimeMs = timeProvider.uptimeInMillis
+        self.appBackgroundTimeMs = timeProvider.millisTime
         logger.log(level: .info, message: "applicationDidEnterBackground", error: nil, data: nil)
     }
 
@@ -142,7 +157,7 @@ final class BaseSessionManager: SessionManager {
     }
 
     private func shouldEndSession() -> Bool {
-        let durationInBackground = timeProvider.uptimeInMillis - appBackgroundTimeMs
+        let durationInBackground = timeProvider.millisTime - appBackgroundTimeMs
 
         if durationInBackground >= configProvider.sessionEndLastEventThresholdMs {
             logger.log(level: .info, message: "Ending session as app was relaunched after being in background for \(durationInBackground) ms", error: nil, data: nil)