+++ ios/MeasureSDKTests/Mocks/MockTimeProvider.swift
@@ -9,25 +9,15 @@ import Foundation
 @testable import MeasureSDK
 
 final class MockTimeProvider: TimeProvider {
-    var elapsedRealtime: Number
-    var currentTimeSinceEpochInMillis: Number
-    var currentTimeSinceEpochInNanos: Number
-    var uptimeInMillis: Number
-    private let iso8601Timestamp: String
+    var millisTime: Number = 0
+    var current: Number = 0
+    var iso8601Timestamp: String = ""
+
+    func now() -> MeasureSDK.Number {
+        return current
+    }
 
     func iso8601Timestamp(timeInMillis: Number) -> String {
         return iso8601Timestamp
-    }
-
-    init(currentTimeSinceEpochInMillis: Number = 0,
-         currentTimeSinceEpochInNanos: Number = 0,
-         uptimeInMillis: Number = 0,
-         iso8601Timestamp: String = "",
-         elapsedRealtime: Number = 0) {
-        self.currentTimeSinceEpochInMillis = currentTimeSinceEpochInMillis
-        self.currentTimeSinceEpochInNanos = currentTimeSinceEpochInNanos
-        self.uptimeInMillis = uptimeInMillis
-        self.iso8601Timestamp = iso8601Timestamp
-        self.elapsedRealtime = elapsedRealtime
     }
 }